fix: bug audit — 8 bugs fixed, zero warnings, all logic tests pass under ASan+UBSan
- game: strict mode no longer counts multiple errors on same char position (key-hammering)
- game: words_typed() returns spaces+1 so word count mode is off-by-one correct
- game: words_typed() counts spaces regardless of correct/wrong (lenient mode fix)
- main: stop timer on all three PLAYING→RESULTS transitions (prevents stale timer events)
- main: restart timer on PLAYING entry (not just at startup)
- main: use uniform_int_distribution instead of rng()%size (eliminates modulo bias)
- main: remove unused sel_file variable
- graphicMgr: remove unused fh in render_menu; fix scroll-down hitbox (was WIN_H+4, now WIN_H-4)
- graphicMgr: all hit-test functions cast mouse coords to float (clean int/float comparisons)
- graphicMgr: draw_passage uses float for fh/fw (eliminates precision loss)
- textbank/game: size_t casts on all vector index sites (zero sign-conversion warnings)

+++ src/textbank/textbank.cpp
@@ -116,7 +116,7 @@ std::string TextBank::load_passage(const FileInfo& fi, Category cat) {
         std::mt19937 rng(std::random_device{}());
         idx = std::uniform_int_distribution<int>(0, (int)paras.size()-1)(rng);
     }
-    std::string passage = paras[idx];
+    std::string passage = paras[(size_t)idx];
     idx++;
     if (idx >= (int)paras.size()) idx = -1; // mark completed
     save_progress();
